Fix lock inversion in python bindings.
When synchronous operations are performed, the python GIL should be
dropped using the `async` macro.  Otherwise, the following deadlock
can occur:
a) A python thread, holding the GIL, may call out to a XrdCl::File
   object, which takes an XRootD-related lock.
b) A XrdCl callback thread, holding an XRootD-related lock, tries
   to invoke a python callback and tries to acquire the GIL.

Since the python GIL is not needed to invoke an XrdCl operation,
this patch drops it for case (a), avoiding the deadlock.